+++ src/test/java/com/greedy/mokkoji/club/service/ClubServiceTest.java
@@ -1,7 +1,7 @@
 package com.greedy.mokkoji.club.service;
 
-import com.greedy.mokkoji.api.club.dto.club.ClubDetailResponse;
-import com.greedy.mokkoji.api.club.dto.club.ClubSearchResponse;
+import com.greedy.mokkoji.api.club.dto.club.response.ClubDetailResponse;
+import com.greedy.mokkoji.api.club.dto.club.response.ClubSearchResponse;
 import com.greedy.mokkoji.api.club.service.ClubService;
 import com.greedy.mokkoji.api.external.AppDataS3Client;
 import com.greedy.mokkoji.db.club.entity.Club;
