+++ src/test/java/com/greedy/mokkoji/favorite/servcie/FavoriteServiceTest.java
@@ -1,6 +1,6 @@
 package com.greedy.mokkoji.favorite.servcie;
 
-import com.greedy.mokkoji.api.club.dto.club.ClubResponse;
+import com.greedy.mokkoji.api.club.dto.club.response.ClubResponse;
 import com.greedy.mokkoji.api.external.AppDataS3Client;
 import com.greedy.mokkoji.api.favorite.service.FavoriteService;
 import com.greedy.mokkoji.common.exception.MokkojiException;
